renamed to bitbots_localization and not a library anymore
