+++ bitbots_localization/cfg/localization.cfg
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-PACKAGE = "humanoid_league_localization"
+PACKAGE = "bitbots_localization"
 
 from dynamic_reconfigure.parameter_generator_catkin import *
