+++ bitbots_localization/include/humanoid_league_localization/tools.h
@@ -1,5 +1,5 @@
-#ifndef HUMANOID_LEAGUE_LOCALIZATION_LOCALIZATION_H
-#define HUMANOID_LEAGUE_LOCALIZATION_LOCALIZATION_H
+#ifndef BITBOTS_LOCALIZATION_LOCALIZATION_H
+#define BITBOTS_LOCALIZATION_LOCALIZATION_H
 
 
 #include <vector>
@@ -12,4 +12,4 @@ std::pair<double, double> cartesianToPolar(double x, double y);
 std::pair<double, double> polarToCartesian(double t, double r);
 
 
-#endif //HUMANOID_LEAGUE_LOCALIZATION_LOCALIZATION_H
+#endif //BITBOTS_LOCALIZATION_LOCALIZATION_H
